requirements.txt
pandas==1.0.3
Flask==1.1.1
numpy==1.19.2
scikit_learn==0.23.2
gunicorn==20.0.4
tensorflow-cpu
keras
scipy==1.4.1
werkzeug
pillow
wheel==0.35
h5py==2.10.0
